requirements.txt
# Core LLM and LangChain dependencies
langgraph==0.3.21
langchain-google-genai==2.1.2
langchain==0.3.18
langchain-core==0.3.51
accelerate==1.10.1

# Local model support
transformers==4.56.0
torch>=2.0.0
python-dotenv==1.0.0

# Audio processing
openai-whisper==20240930
piper-tts==1.2.0
soundfile==0.13.1

# UI
gradio==5.25.0

# Utilities
numpy==1.26.4